+++ app/src/main/java/com/example/bp3/MainActivity.java
@@ -6,12 +6,13 @@ import android.support.v4.app.FragmentTransaction;
 import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
-import com.example.bp3.utils.helpers.RestApiHelper;
+
+import com.example.bp3.service.repository.RestApiHelper;
 import com.example.bp3.views.fragments.ChallengeView;
 import com.example.bp3.views.fragments.MyStuffView;
 import com.example.bp3.views.fragments.OpdrachtView;
@@ -39,6 +40,8 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
 
         //DIT WORDT HET HOMESCREEN
 //        displaySelectedScreen(R.id.navigation_assignments);
+
+
     }
 
     @Override
