+++ app/src/main/java/com/example/bp3/service/models/Activiteit.java
@@ -1,17 +1,12 @@
-package com.example.bp3.models;
-import android.util.Log;
+package com.example.bp3.service.models;
 
 import com.google.gson.annotations.SerializedName;
 
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
-import java.time.OffsetDateTime;
-import java.time.ZoneOffset;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Date;
 
-import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
