+++ app/src/main/java/com/example/bp3/service/models/Bedrijf.java
@@ -1,19 +1,12 @@
-package com.example.bp3.models;
-
-import com.google.gson.annotations.SerializedName;
-
-import java.util.ArrayList;
+package com.example.bp3.service.models;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class Docent extends Account {
-    @SerializedName("opleiding")
-    private Opleiding opleiding;
-    @SerializedName("lesvakCollection")
-    private ArrayList<Lesvak> lesvakken;
+public class Bedrijf extends Account {
+    private String email_contactpersoon, telefoon_contactpersoon;
 
     public boolean login (String email, String wachtwoord){
         //controleer email en wachtwoord combinatie
