+++ app/src/main/java/com/example/bp3/service/models/Challenge.java
@@ -1,4 +1,4 @@
-package com.example.bp3.models;
+package com.example.bp3.service.models;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
