+++ app/src/main/java/com/example/bp3/viewmodels/OpdrachtViewModel.java
@@ -1,0 +1,9 @@
+package com.example.bp3.viewmodels;
+
+public class OpdrachtViewModel {
+
+    public OpdrachtViewModel() { //Account account in params
+
+
+    }
+}
