+++ app/src/main/java/com/example/bp3/views/adapter/CustomBindingAdapter.java
@@ -1,0 +1,12 @@
+package com.example.bp3.views.adapter;
+
+import android.databinding.BindingAdapter;
+import android.view.View;
+
+
+public class CustomBindingAdapter {
+    @BindingAdapter("visibleGone")
+    public static void showHide(View view, boolean show) {
+        view.setVisibility(show ? View.VISIBLE : View.GONE);
+    }
+}
